requirements.txt
# Production dependencies
fastapi
uvicorn[standard]
apify-client
python-picnic-api
pydantic
python-dotenv
requests

# Frontend dependencies (Streamlit)
streamlit
pandas

# Dev dependencies (for testing)
pytest
